+++ src/sis/section/@types.ts
@@ -1,0 +1,14 @@
+import { WeekDay } from "../../generated/prisma";
+
+export enum SectionCourseScheduleCreationStatus {
+    UnknownSection,
+    UnknownCourse,
+    UnknownFaculty,
+    Success,
+}
+
+export interface SectionCourseScheduleOptions {
+    weekDay: WeekDay;
+    startTime: string;
+    endTime: string;
+}