+++ src/sis/section/sections.ts
@@ -2,7 +2,15 @@ import { z } from "zod";
 
 import prisma, { Prisma } from "../../services/prisma";
 
+import {
+    SectionCourseScheduleCreationStatus,
+    SectionCourseScheduleOptions,
+} from "./@types";
+
 import { programs } from "../program";
+import { inclusions, User, UserName, users } from "../../user";
+import { courses } from "../course";
+import { GradeStatus, WeekDay } from "../../generated/prisma";
 
 class SectionManager {
     protected _dataSchema = z.object({
@@ -13,12 +21,35 @@ class SectionManager {
 
     protected _partialDataSchema = this._dataSchema.partial();
 
+    protected _scheduleSchema = z.object({
+        code: z.string(),
+        course: z.string(),
+        faculty: z.object({
+            given: z.string(),
+            middle: z.string().optional(),
+            last: z.string(),
+        }),
+        scheduleSlots: z.array(
+            z.object({
+                weekDay: z.enum(WeekDay),
+                startTime: z
+                    .string()
+                    .regex(/^(0?[1-9]|1[0-2]):[0-5]\d (AM|PM)$/),
+                endTime: z.string().regex(/^(0?[1-9]|1[0-2]):[0-5]\d (AM|PM)$/),
+            })
+        ),
+    });
+
     public get schema() {
         return this._dataSchema;
     }
 
     public get partialSchema() {
         return this._partialDataSchema;
+    }
+
+    public get scheduleSchema() {
+        return this._scheduleSchema;
     }
 
     public async create(programCode: string, code: string, year: number) {
@@ -35,10 +66,14 @@ class SectionManager {
     }
 
     public async update(code: string, data: Prisma.SectionUpdateInput) {
-        return (
-            (await prisma.section.updateMany({ where: { code }, data })).count >
-            0
-        );
+        try {
+            return (
+                await prisma.section.update({ where: { code }, data }),
+                true
+            );
+        } catch {
+            return false;
+        }
     }
 
     public async delete(code: string) {
@@ -50,14 +85,87 @@ class SectionManager {
     }
 
     public async get(code: string) {
-        return await prisma.section.findUnique({ where: { code } });
+        return await prisma.section.findUnique({
+            where: { code },
+            include: inclusions.section,
+        });
     }
 
     public async getAllFrom(programCode: string) {
         const program = await programs.get(programCode);
         if (!program) return [];
 
-        return await prisma.section.findMany({ where: { program } });
+        return await prisma.section.findMany({
+            where: { program },
+            include: inclusions.section,
+        });
+    }
+
+    /**
+     * @param code The section code of the section to create the schedule for
+     * @param course The course to create the schedule for the given section
+     * @param faculty The The faculty/professor to be assigned to the course and section
+     */
+    public async createSchedule(
+        code: string,
+        course: string,
+        faculty: UserName,
+        scheduleSlots: SectionCourseScheduleOptions[]
+    ): Promise<SectionCourseScheduleCreationStatus> {
+        const section = await this.get(code);
+        if (!section) return SectionCourseScheduleCreationStatus.UnknownSection;
+
+        const coursePayload = await courses.get(course);
+        if (!coursePayload)
+            return SectionCourseScheduleCreationStatus.UnknownCourse;
+
+        const facultyUser = await users.getByName(faculty);
+        await facultyUser?.initialize();
+
+        if (!facultyUser || !facultyUser.faculty)
+            return SectionCourseScheduleCreationStatus.UnknownFaculty;
+
+        await prisma.courseSchedule.create({
+            data: {
+                section: { connect: { id: section.id } },
+                course: { connect: { id: coursePayload.id } },
+                faculty: { connect: { id: facultyUser.faculty.id } },
+
+                schedule: { create: scheduleSlots },
+            },
+        });
+
+        return SectionCourseScheduleCreationStatus.Success;
+    }
+
+    public async assign(code: string, user: User) {
+        if (!user.student) return;
+
+        const section = await this.get(code);
+        console.log(section);
+        if (!section) return;
+
+        await this.update(code, { students: { connect: { id: user.id } } });
+
+        try {
+            console.log(section.courses);
+            await prisma.studentUser.update({
+                where: { userId: user.id },
+                data: {
+                    courses: {
+                        create: section.courses.map((sched) => {
+                            return {
+                                courseSched: { connect: { id: sched.id } },
+                                gradeStatus: GradeStatus.Pending,
+                                grade: 0.0,
+                            };
+                        }),
+                    },
+                },
+            });
+        } catch (err) {
+            console.log(err);
+        }
     }
 }